requirements.txt
astroid==2.2.5
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pylint==2.3.1
six==1.12.0
steg==0.1.2
typed-ast==1.3.1
wrapt==1.11.1
